Implement extra pipeline for predicting using the best model
This pipeline is implemented as an exercise. It is composed of 3 blocks,
responsible for loading the best model from mlflow, using it to make a
predictions on the received payload, and exporting the prediction with
the input data to a file. The pipeline has an API trigger implemented,
so that it can accept external requests. It is also monitored with a
default dashboard.

Note that: the "best model" from mlflow is a pointer, not a fixed
object. This is currently implemented as "the model with the lowest
RMSE", but it may well use a registered/aliased model if that's wanted.

Note also that exporting the predictions (to a file in this case) serves
monitoring purposes. For example, it would be easy to implement a chart
in the dashboard that shows the distribution of the historical
predictions.